Fix all compilation errors - desktop client builds successfully
Fixed multiple compilation issues to get desktop client building:

macOS Input Simulation:
- Removed CGEventSource from struct (not Send-safe)
- Create event source on-demand in each function
- Fixed scroll event to use EventField enum
- Fixed CGPoint usage in mouse events

macOS Screen Capture:
- Removed unavailable CGDisplay APIs
- Fixed data.len() type casting (isize → usize)

Network Layer:
- Fixed ConnectionManager::new() - removed incorrect await
- Added clone() to ConnectionState::get_state()

File Transfer:
- Fixed borrow checker issues in calculate_checksum
- Properly drop MutexGuard before calling mutable methods

All modules now compile successfully with only warnings.
Ready for testing and release builds.

+++ scrdesk/client/desktop/src/capture/macos.rs
@@ -1,6 +1,6 @@
 use super::{Frame, ScreenCapture};
 use anyhow::{Context, Result};
-use core_graphics::display::{CGDisplay, CGDisplayStreamRef, kCGDisplayStreamShowCursor};
+use core_graphics::display::CGDisplay;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 pub struct MacOSCapturer {
@@ -58,7 +58,7 @@ impl ScreenCapture for MacOSCapturer {
         for y in 0..height {
             for x in 0..width {
                 let offset = (y as usize * bytes_per_row) + (x as usize * bytes_per_pixel);
-                if offset + 3 < data.len() {
+                if offset + 3 < data.len() as usize {
                     rgba_data.push(data[offset + 2]); // R
                     rgba_data.push(data[offset + 1]); // G
                     rgba_data.push(data[offset]);     // B
